requirements.txt
mastodoner
pytest
datasets
setfit
sentence-transformers
psycopg2
python-dotenv
beautifulsoup4
lxml
emoji
torch 
transformers 
bitsandbytes 
tqdm
huggingface_hub
bitsandbytes
optuna
scikit-learn
py3langid
matplotlib